Fix for lmfit nan-policy, merged *_star.py files with *.py
