+++ descwl/__init__.py
@@ -11,8 +11,3 @@ import render
 import analysis
 import output
 import trace
-import model_star
-import render_star
-import analysis_star
-import catalog_star
-import output_star
